Rename pkg package to mandelbrot
