+++ mandelbrot/mandelbrot.go
@@ -1,4 +1,4 @@
-package pkg
+package mandelbrot
 
 import (
 	"fmt"
